# user-address-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Address API",
        "description": "API para gerenciamento de usuários e seus contatos com endereços",
        "contact": {
            "email": "admin@useraddress.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "User Address API Server"
        }
    ],
    "paths": {
...
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Listar contatos do usuário",
                "description": "Retorna todos os contatos do usuário autenticado com filtros opcionais",
                "operationId": "0f9efb5609dbb039b35ecafb28e492a2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nome, email, telefone, CPF ou endereço",
                        "schema": {
                            "type": "string",
                            "example": "João"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtrar por cidade",
                        "schema": {
                            "type": "string",
                            "example": "São Paulo"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "example": "SP"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo para ordenação",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "email",
                                "city",
                                "created_at"
                            ],
                            "example": "name"
                        }
                    },
                    {
                        "name": "order_direction",
                        "in": "query",
                        "description": "Direção da ordenação",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (máximo 100)",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contatos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Contact"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 68
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Criar novo contato",
                "description": "Criar um novo contato para o usuário autenticado",
                "operationId": "a3b0baf957a16c1ed505f2e8985c254c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "cpf",
                                    "cep",
                                    "street",
                                    "number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "01234-567"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Rua das Flores"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "example": "Apto 101"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contato criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contato criado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Contact"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contact": {
                "title": "Contact",
                "description": "Modelo de contato",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Maria Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "maria@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "11999999999"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123.456.789-00"
                    },
                    "cep": {
                        "type": "string",
                        "example": "01234-567"
                    },
                    "street": {
                        "type": "string",
                        "example": "Rua das Flores"
                    },
                    "number": {
                        "type": "string",
                        "example": "123"
                    },
                    "complement": {
                        "type": "string",
                        "example": "Apto 101"
                    },
                    "neighborhood": {
                        "type": "string",
                        "example": "Centro"
                    },
                    "city": {
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "state": {
                        "type": "string",
                        "example": "SP"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -23.5505
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -46.6333
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}